Split Note Scheduler into smaller classes
