+++ src/audio/engine/WebAudioEngine.js
@@ -2,6 +2,7 @@ import NoteScheduler from './NoteScheduler';
 import sineTestNote from 'audio/notes';
 
 const AUDIO_CONTEXT = 'audioContext';
+export const DEFAULT_NOTE_LENGTH = 0.5;
 
 class WebAudioEngine {
